my_library/requirements.txt
Flask
mysql-connector-python
openai
textwrap3
werkzeug
